Implmented solution to 0 our rows and columns in 2d array if the row/column contains a zero
note: this problem requested the matrix be modified
in place on leetcode.

This performs two passes over the matrix in each direction
generating lists of row/column indexes that need to be zeroed
out. A bit of room for improvement here as we could probally
generate these lists in a single pass.

Then we do a single pass over the rows in the matrix,
updating them all to zeros by copying in an empty row.
Or make a pass over the columns that need to be zeroed
out.